Add test report functionality
Provided the reporter as cypress multi reporters and provided the
config file, which is responsible for the configuration of the
reports in the report config to use mochawesome-unit-reporter and
junit reporter.
Running the 'cypress:e2e' cmd will start the server, run all tests,
generate reports, then merge the reports in the folder paths specified.